Initial project revision.
This package was originally part of the Mentat system (https://mentat.cesnet.cz). It started to be usefull for other projects as well, so we decided to move it to separate git repository. This initial revision was created by moving the relevant source files from Mentat
repository revision rev:5f76179.