Fixed issue with timestamps and timezones
A test failure occured during a build in copr (there is probably
different timezone).  After that, I discovered a strange behaviour when
TZ is changed.  Timestamp from IDEA is probably loaded or converted in
some bad way, because the difference between expected and computed time
changes with timezones.

As a workaround, expected (float) constant was replaced by expected object
created by summing seconds and the timestamp in the
test_05_advanced_filters() test.

In addition, the source code was modified to be compatible with python2
(datetime has no timestamp() there).