Added special handling of IN operator with list of IP subnet
test_filters_idea.py now contains one new test where Source.IP4 is
compared with a list with IP4Net.  In such case, it is not possible to
use lambda function `x in y` because it evaluates as an exact match.

This is a patch that improves RuleTreeTraverser in order to support this
kind of lists (all tests pass), however, it may not be an ideal solution...

Therefore, I'm asking for review/hint.