Rewritten special handling of IN operator for lists of IP ranges.
The 80c07d10878d360539def0db9da14fca493211ce commit fixed the bug in comparing IP address objects using IN operator with list of IP addresses or networks. The original solution simply did not work and did not return True, unless there was exact match. The solution used to resolve this problem worked correcty, it did not however meet the philosophy of this library. It introduced unnecessary complexity and advanced logic to lower layers of the library and on top of that this logic was executed again and again for each processed message/object.

This patch attempts for a different approach using already introduced concept of rule tree compilation. This compilation is performed exactly once for each rule tree and it modifies and optimizes the tree structure and replaces certain tree objects to achieve the same result.

As a result of this change some parts of the code had to be moved from rules.py module to filters.py module. This was necessary to prevent circular references in module imports, however this ordering seems more appropriate anyway.

All current unit tests pass, so this patch should not break anything.