Fix: Finally got rid of annoying "Token 'EXP_ALL' defined, but not used" parser warning message.
The pynspect.lexer uses EXP_ALL token to consolidate regular expressions for all operators into one huge regular expression. This is according to the recomendations in official documentation. However this token is only virtual as is not used anywhere in the grammar and parser is complaning about that. I have finally found a tip how to get rid of this message, because it spams terminal in all projects, that are using this library.